+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Fix: [issue 218](https://github.com/rubiii/savon/pull/218) - Savon now correctly handles namespaced
   Array items in a Hash passed to `Savon::SOAP::XML#body=`.
 
+* Improvement: [issue 222](https://github.com/rubiii/savon/issues/222) - Set the Content-Length header.
+
 ## 0.9.6 (2011-07-07)
 
 * Improvement/Fix: Updated Savon to use the latest version of [Wasabi](http://rubygems.org/gems/wasabi).